requirements.txt
python-dotenv==1.0.1 # For reading environment variables stored in .env file
langchain
langchain-community
langchain-openai # For embeddings
unstructured==0.14.4 # Document loading
# onnxruntime==1.17.1 # chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# install onnxruntime before installing `chromadb`
chromadb==0.5.0 # Vector storage
openai # For embeddings
tiktoken==0.7.0  # For embeddings 
boto3==1.34.143
pytest==8.2.2
streamlit
# install markdown depenendies with: `pip install "unstructured[md]"` after install the requirements file. Leave this line commented out. 